Refactorize schemas for restructuring of application
