+++ .eslintrc.js
@@ -3,6 +3,7 @@ module.exports = {
     browser: true,
     commonjs: true,
     es6: true,
+    mocha: true,
   },
   extends: 'airbnb-base',
   globals: {
