+++ components/person/schema.js
@@ -1,5 +1,5 @@
 const Joi = require('@hapi/joi');
-const globalSchema = require('./globalSchema');
+const globalSchema = require('../validation/global.schema');
 
 // Build your endpoint-specific schema here.
 const body = Joi.object().keys({
