--- components/validation/testSchema.js
@@ -1,11 +1,0 @@
-const Joi = require('@hapi/joi');
-const globalSchema = require('./globalSchema');
-
-// Build your endpoint-specific schema here.
-const body = Joi.object().keys({
-  id: globalSchema.id,
-});
-
-module.exports = {
-  body,
-};
