+++ package.json
@@ -4,7 +4,9 @@
   "description": "Boilerplate repository for nodejs based LABS projects.",
   "main": "app.js",
   "scripts": {
-    "start": "nodemon -r dotenv/config app.js",
+    "start": "NODE_ENV=production node -r dotenv/config app.js",
+    "dev": "nodemon -r dotenv/config app.js",
+    "test": "NODE_ENV=test ./node_modules/.bin/mocha components/**/*.test.js",
     "test:watch": "./node_modules/.bin/mocha components/**/*.test.js --watch"
   },
   "author": "Helsingborg Stad",
@@ -27,6 +29,7 @@
     "dotenv": "^8.0.0",
     "easy-soap-request": "^2.5.0",
     "express": "^4.16.4",
+    "lodash": "^4.17.11",
     "redis": "^2.8.0",
     "soap": "^0.27.1",
     "swagger-ui-express": "^4.0.2"
